More view layer renaming
